Remove symlinks
Change the way we include the Vagrantfile to avoid using symlinks.

Symlinks do not work on Windows, so instead we put the configuration
in a Vagrantfile that lies in the project directory which then loads
the "real" Vagrantfile from the submodule.

Thanks @chregu for the initial PR and the idea !