Cleanup CustomConfig and default parameter values
New rule of thumbs for paramters :

* If it is common for most project -> Drifter Vagrantfile
* If it is project specific -> parameters.yaml
* Some calculation into the project Vagrantfile

I kept the 'ansible_local' parameter only in the project
Vagrantfile because most people shouldn't have to change it
but we still want to allow it if need be.

Also, the mess around lxc_* and vbox_* url and name has been
removed in favor of using the JSON based box format.