Improve MySQL role and use upstream repository
Switch to the MySQL upstream repository which supports
multiple and newer versions of MySQL. The version can now
be specified by the parameter `mysql_version` and defaults
to 5.6.

When the version changes, the role will uninstall the currently
installed version of MySQL. This way the switch from the distribution
package to the upstream package should work without problems.